Move expected values
