+++ src/pykeen/evaluation/rank_based_evaluator.py
@@ -9,20 +9,7 @@ import random
 import re
 from collections import defaultdict
 from dataclasses import fields
-from typing import (
-    Any,
-    Iterable,
-    List,
-    Mapping,
-    MutableMapping,
-    NamedTuple,
-    Optional,
-    Sequence,
-    Tuple,
-    Type,
-    Union,
-    cast,
-)
+from typing import Any, Iterable, List, Mapping, MutableMapping, NamedTuple, Optional, Tuple, Type, Union, cast
 
 import numpy as np
 import pandas as pd
@@ -525,24 +512,3 @@ class SampledRankBasedEvaluator(RankBasedEvaluator):
         # write back correct num_entities
         # TODO: should we give num_entities in the constructor instead of inferring it every time ranks are processed?
         self.num_entities = all_scores.shape[1]
-
-
-def numeric_expected_value(
-    metric: str,
-    num_candidates: Union[Sequence[int], np.ndarray],
-    num_samples: int,
-) -> float:
-    """
-    Compute expected metric value by summation.
-
-    Depending on the metric, the estimate may not be very accurate and converage slowly, cf.
-    https://docs.scipy.org/doc/scipy/reference/generated/scipy.stats.rv_discrete.expect.html
-    """
-    metric_func = metric_resolver.make(metric)
-    num_candidates = np.asarray(num_candidates)
-    generator = np.random.default_rng()
-    expectation = 0
-    for _ in range(num_samples):
-        ranks = generator.integers(low=0, high=num_candidates)
-        expectation += metric_func(ranks)
-    return expectation / num_samples
